Move some code from shared to main/worker side in the bundle (internal-1302)
* move worker tile code to worker side in the bundle

* remove source cache from shared bundle

* move validation code from shared to main

* fix unit tests

* fix lint

* move transform code to index

* remove outdated tests

* fixup validate options

* fix validate condition again
