+++ src/data/bucket/line_bucket.js
@@ -15,6 +15,8 @@ import {hasPattern, addPatternDependencies} from './pattern_bucket_features.js';
 import loadGeometry from '../load_geometry.js';
 import toEvaluationFeature from '../evaluation_feature.js';
 import EvaluationParameters from '../../style/evaluation_parameters.js';
+import '../../render/line_atlas.js'; // make sure LineAtlas is registered for deserialization
+
 import type {ProjectionSpecification} from '../../style-spec/types.js';
 import type {CanonicalTileID} from '../../source/tile_id.js';
 import type {
