+++ src/source/vector_tile_worker_source.js
@@ -1,7 +1,5 @@
 // @flow
 
-import {getArrayBuffer} from '../util/ajax.js';
-
 import {VectorTile} from '@mapbox/vector-tile';
 import Protobuf from 'pbf';
 import WorkerTile from './worker_tile.js';
@@ -9,127 +7,20 @@ import {extend} from '../util/util.js';
 import {getPerformanceMeasurement} from '../util/performance.js';
 import {Evented} from '../util/evented.js';
 import tileTransform from '../geo/projection/tile_transform.js';
+import {loadVectorTile, DedupedRequest} from './load_vector_tile.js';
 
 import type {
     WorkerSource,
     WorkerTileResult,
     WorkerTileParameters,
-    RequestedTileParameters,
     WorkerTileCallback,
     TileParameters
 } from './worker_source.js';
 
 import type Actor from '../util/actor.js';
 import type StyleLayerIndex from '../style/style_layer_index.js';
-import type {Callback} from '../types/callback.js';
 import type Scheduler from '../util/scheduler.js';
-import type {IVectorTile} from '@mapbox/vector-tile';
-
-export type LoadVectorTileResult = {
-    rawData: ArrayBuffer;
-    vectorTile?: IVectorTile;
-    expires?: any;
-    cacheControl?: any;
-    resourceTiming?: Array<PerformanceResourceTiming>;
-};
-
-/**
- * @callback LoadVectorDataCallback
- * @param error
- * @param vectorTile
- * @private
- */
-export type LoadVectorDataCallback = Callback<?LoadVectorTileResult>;
-
-export type AbortVectorData = () => void;
-export type LoadVectorData = (params: RequestedTileParameters, callback: LoadVectorDataCallback) => ?AbortVectorData;
-export class DedupedRequest {
-    entries: { [string]: Object };
-    scheduler: ?Scheduler;
-
-    constructor(scheduler?: Scheduler) {
-        this.entries = {};
-        this.scheduler = scheduler;
-    }
-
-    request(key: string, metadata: Object, request: any, callback: LoadVectorDataCallback): (() => void) {
-        const entry = this.entries[key] = this.entries[key] || {callbacks: []};
-
-        if (entry.result) {
-            const [err, result] = entry.result;
-            if (this.scheduler) {
-                this.scheduler.add(() => {
-                    callback(err, result);
-                }, metadata);
-            } else {
-                callback(err, result);
-            }
-            return () => {};
-        }
-
-        entry.callbacks.push(callback);
-
-        if (!entry.cancel) {
-            entry.cancel = request((err, result) => {
-                entry.result = [err, result];
-                for (const cb of entry.callbacks) {
-                    if (this.scheduler) {
-                        this.scheduler.add(() => {
-                            cb(err, result);
-                        }, metadata);
-                    } else {
-                        cb(err, result);
-                    }
-                }
-                setTimeout(() => delete this.entries[key], 1000 * 3);
-            });
-        }
-
-        return () => {
-            if (entry.result) return;
-            entry.callbacks = entry.callbacks.filter(cb => cb !== callback);
-            if (!entry.callbacks.length) {
-                entry.cancel();
-                delete this.entries[key];
-            }
-        };
-    }
-}
-
-/**
- * @private
- */
-// $FlowFixMe[missing-this-annot]
-export function loadVectorTile(params: RequestedTileParameters, callback: LoadVectorDataCallback, skipParse?: boolean): (() => void) {
-    const key = JSON.stringify(params.request);
-
-    const makeRequest = (callback: LoadVectorDataCallback) => {
-        const request = getArrayBuffer(params.request, (err: ?Error, data: ?ArrayBuffer, cacheControl: ?string, expires: ?string) => {
-            if (err) {
-                callback(err);
-            } else if (data) {
-                callback(null, {
-                    vectorTile: skipParse ? undefined : new VectorTile(new Protobuf(data)),
-                    rawData: data,
-                    cacheControl,
-                    expires
-                });
-            }
-        });
-        return () => {
-            request.cancel();
-            callback();
-        };
-    };
-
-    if (params.data) {
-        // if we already got the result earlier (on the main thread), return it directly
-        (this.deduped: DedupedRequest).entries[key] = {result: [null, params.data]};
-    }
-
-    const callbackMetadata = {type: 'parseTile', isSymbolTile: params.isSymbolTile, zoom: params.tileZoom};
-    return (this.deduped: DedupedRequest).request(key, callbackMetadata, makeRequest, callback);
-}
+import type {LoadVectorData} from './load_vector_tile.js';
 
 /**
  * The {@link WorkerSource} implementation that supports {@link VectorTileSource}.
