+++ src/style/style_layer/raster_particle_style_layer.js
@@ -6,7 +6,6 @@ import properties from './raster_particle_style_layer_properties.js';
 import {PossiblyEvaluated} from '../properties.js';
 import {renderColorRamp} from '../../util/color_ramp.js';
 import {RGBAImage} from '../../util/image.js';
-import SourceCache from '../../source/source_cache.js';
 
 import type {ConfigOptions} from "../properties";
 import type MapboxMap from '../../ui/map.js';
@@ -14,6 +13,7 @@ import type {PaintProps} from './raster_particle_style_layer_properties.js';
 import type {LayerSpecification} from '../../style-spec/types.js';
 import type Texture from '../../render/texture.js';
 import type Framebuffer from '../../gl/framebuffer.js';
+import type SourceCache from '../../source/source_cache.js';
 
 const COLOR_RAMP_RES = 256;
